Origin/feature/yje/card

+++ src/card/card.service.ts
@@ -1,28 +1,52 @@
-import { Injectable, BadRequestException } from '@nestjs/common';
+import {
+  Injectable,
+  BadRequestException,
+  NotFoundException,
+} from '@nestjs/common'; // NestJS의 기본 모듈을 가져옵니다.
 import { CreateCardDto } from './dto/create-card.dto';
 import { UpdateCardDto } from './dto/update-card.dto';
-import { Repository } from 'typeorm';
+import { Repository } from 'typeorm'; // TypeORM의 Repository를 가져옵니다.
 import { InjectRepository } from '@nestjs/typeorm';
 import { Card } from './entities/card.entity';
+import { Member } from 'src/member/entities/member.entity';
+import { Board } from 'src/board/entities/board.entity';
 import { LexoRank } from 'lexorank';
 import _ from 'lodash';
+import { CreateColumnDto } from 'src/column/dto/create-column.dto';
+import { ColumnEntity } from 'src/column/entities/column.entity';
 
-// 카드가 만들어지는데에 필요한 정보를 정해준다
-// 타입 orm을 써서 DB에 저장한다 (받아온 정보로 카드를 만든다)
-// 저장한 정보를 리턴한다
-
+// 카드 서비스를 정의하는 클래스입니다.
 @Injectable()
 export class CardService {
-  // 컨스트럭터 작성
-  // 컨스트럭터 안에 레포지토리 연결하기
-  // 크리에이트 매서드 생성
-  // 파람스 해서 타입지정?
-
+  // 생성자에서 카드 레포지토리를  주입받습니다.
   constructor(
-    @InjectRepository(Card)
-    private cardRepository: Repository<Card>,
+    @InjectRepository(ColumnEntity)
+    private columnRepository: Repository<ColumnEntity>,
+    @InjectRepository(Member) private memberRepository: Repository<Member>,
+    @InjectRepository(Board) private boardRepository: Repository<Board>,
+    @InjectRepository(Card) // Card 엔티티에 대한 레포지토리를 주입합니다.
+    private cardRepository: Repository<Card>, // 주입된 레포지토리를 private 변수로 선언합니다.
   ) {}
-  async createCard(columnId: number, createCardDto: CreateCardDto) {
+  async createCard(
+    userId: number,
+    columnId: number,
+    createCardDto: CreateCardDto,
+  ) {
+    const column = await this.columnRepository.findOne({
+      where: { id: columnId },
+      relations: ['board'],
+    });
+
+    const checkMember = await this.memberRepository.findOne({
+      where: { userId: userId, boardId: column.board.id },
+    });
+
+    if (!checkMember) {
+      throw new NotFoundException(
+        '카드을 만들수 있는 권한이 존재하지 않습니다.',
+      );
+    }
+
     let lexoRank: LexoRank;
     const existingCard = await this.cardRepository.findOne({
       where: {},
@@ -37,7 +61,7 @@ export class CardService {
     const newCard: Card = this.cardRepository.create({
       title: createCardDto.title,
       content: createCardDto.content,
-      memberId: createCardDto.memberId,
+      memberId: checkMember.id, // 멤버아이디
       lexo: lexoRank.toString(),
       columnId: columnId,
     });
@@ -46,28 +70,83 @@ export class CardService {
     return savedCard;
   }
 
-  async findCards(columnId: number): Promise<Card[]> {
+  // 특정 열에 속한 카드 목록을 찾는 메서드
+
+  async findCards(userId: number, columnId: number): Promise<Card[]> {
+    const column = await this.columnRepository.findOne({
+      where: { id: columnId },
+      relations: ['board'],
+    });
+
+    const checkMember = await this.memberRepository.findOne({
+      where: { userId: userId, boardId: column.board.id },
+    });
+
+    if (!checkMember) {
+      throw new NotFoundException(
+        '카드을 조회할 수 있는 권한이 존재하지 않습니다.',
+      );
+    }
+
+    // 주어진 열 ID에 따라 모든 카드를 조회하여 반환합니다.
     return await this.cardRepository.find({
-      where: { columnId },
+      where: { columnId }, // 조건: columnId
     });
   }
 
-  async findCard(columnId: number, cardId: number): Promise<Card> {
-    const card = await this.cardRepository.findOne({
-      where: { id: cardId, columnId },
+  // 특정 카드 하나를 찾는 메서드
+  async findCard(
+    userId: number,
+    columnId: number,
+    cardId: number,
+  ): Promise<Card> {
+    const column = await this.columnRepository.findOne({
+      where: { id: columnId },
+      relations: ['board'],
     });
+
+    const checkMember = await this.memberRepository.findOne({
+      where: { userId: userId, boardId: column.board.id },
+    });
+
+    if (!checkMember) {
+      throw new NotFoundException(
+        '카드을 조회할 수 있는 권한이 존재하지 않습니다.',
+      );
+    }
+
+    const card = await this.cardRepository.findOne({
+      where: { id: cardId, columnId }, // 카드 ID와 열 ID로 카드 조회
+    });
+    // 카드가 존재하지 않을 경우 예외를 발생시킵니다.
     if (!card) {
       throw new BadRequestException('카드가 존재하지 않습니다.');
     }
 
-    return card;
+    return card; // 카드가 존재하면 반환합니다.
   }
 
   async updateCardOrder(
+    userId: number,
     columnId: number,
     cardId: number,
     targetCardId: number,
   ): Promise<Card> {
+    const column = await this.columnRepository.findOne({
+      where: { id: columnId },
+      relations: ['board'],
+    });
+
+    const checkMember = await this.memberRepository.findOne({
+      where: { userId: userId, boardId: column.board.id },
+    });
+
+    if (!checkMember) {
+      throw new NotFoundException(
+        '카드을 변경할 수 있는 권한이 존재하지 않습니다.',
+      );
+    }
+
     const cards = await this.cardRepository.find({
       where: { columnId },
       order: { lexo: 'DESC' },
@@ -86,13 +165,11 @@ export class CardService {
     if (targetCardIndex < CardIndex) {
       const targetNextCardIndex = targetCardIndex - 1;
 
-
       if (!card || !targetCard) {
         throw new BadRequestException('카드가 존재하지 않습니다.');
       }
 
       const currentRank = LexoRank.parse(card.lexo);
-      console.log(`--------------------> 현재 카드 랭크`, currentRank);
       const targetRank = LexoRank.parse(targetCard.lexo);
       const targetNextCard = cards[targetNextCardIndex];
 
@@ -139,7 +216,6 @@ export class CardService {
       }
 
       const currentRank = LexoRank.parse(card.lexo);
-      console.log(`--------------------> 현재 카드 랭크`, currentRank);
       const targetRank = LexoRank.parse(targetCard.lexo);
       const targetNextCard = cards[targetNextCardIndex];
 
@@ -154,7 +230,6 @@ export class CardService {
 
         return await this.cardRepository.findOne({ where: { id: cardId } });
       }
-      console.log('----------------------------');
       const newRank = LexoRank.parse(targetNextCard.lexo).between(targetRank); // 현재 카드와 타켓 카드 사이의 랭크
 
       await this.cardRepository.update(
@@ -167,12 +242,33 @@ export class CardService {
     // 카드의 순서 업데이트
   }
 
-  async deleteCard(cardId: number, columnId: number): Promise<void> {
-    const result = await this.cardRepository.delete({
-      id: cardId,
-      columnId,
+  // 카드 삭제 메서드
+  async deleteCard(
+    userId: number,
+    cardId: number,
+    columnId: number,
+  ): Promise<void> {
+    const column = await this.columnRepository.findOne({
+      where: { id: columnId },
+      relations: ['board'],
     });
 
+    const checkMember = await this.memberRepository.findOne({
+      where: { userId: userId, boardId: column.board.id },
+    });
+
+    if (!checkMember) {
+      throw new NotFoundException(
+        '카드을 삭제할 수 있는 권한이 존재하지 않습니다.',
+      );
+    }
+    // 주어진 카드 ID와 열 ID에 해당하는 카드를 삭제합니다.
+    const result = await this.cardRepository.delete({
+      id: cardId,
+      columnId, // 삭제 조건: columnId
+    });
+
+    // 삭제된 카드가 없는 경우 예외를 발생시킵니다.
     if (result.affected === 0) {
       throw new BadRequestException('카드를 찾지 못했습니다.');
     }
